requirements.txt
httpx==0.25.2
python-dotenv==1.0.0